Lower media weight if catalognum exists
The release distance isn't lowered if it doesn't have a catalog number (as far as I can tell).
If the media is removed, then if one release matches both the media and catalog number, and another has a non-matching media and doesn't have a catalog number, then the non-matching media wouldn't be penalized.

For examples, this plugin chooses the cassette release for this album instead of the CD, even though both media and catalog number match in origin.yaml
https://musicbrainz.org/release-group/05c73df7-423f-4dab-b600-afdc7d453e4e
Instead, it might be good to lower the media weight.